Separate Farnocchia propagator into separate module
