+++ src/poliastro/core/propagation/__init__.py
@@ -7,15 +7,25 @@ from ..angles import (
     E_to_nu,
     F_to_M,
     F_to_nu,
-    M_to_nu,
     _kepler_equation,
     _kepler_equation_prime,
-    nu_to_M,
     nu_to_E,
     nu_to_F,
 )
 from ..elements import coe2rv, rv2coe
 from ..stumpff import c2, c3
+from .farnocchia import farnocchia
+
+__all__ = [
+    "func_twobody",
+    "farnocchia",
+    "vallado",
+    "mikkola",
+    "markley",
+    "pimienta",
+    "gooding",
+    "danby",
+]
 
 
 def func_twobody(t0, u_, k, ad, ad_kwargs):
@@ -44,58 +54,6 @@ def func_twobody(t0, u_, k, ad, ad_kwargs):
 
     du = np.array([vx, vy, vz, -k * x / r3 + ax, -k * y / r3 + ay, -k * z / r3 + az])
     return du
-
-
-@jit
-def farnocchia(k, r0, v0, tof):
-    r"""Propagates orbit using mean motion.
-
-    This algorithm depends on the geometric shape of the orbit.
-    For the case of the strong elliptic or strong hyperbolic orbits:
-
-    ..  math::
-
-        M = M_{0} + \frac{\mu^{2}}{h^{3}}\left ( 1 -e^{2}\right )^{\frac{3}{2}}t
-
-    .. versionadded:: 0.9.0
-
-    Parameters
-    ----------
-    k : float
-        Standar Gravitational parameter
-    r0 : ~astropy.units.Quantity
-        Initial position vector wrt attractor center.
-    v0 : ~astropy.units.Quantity
-        Initial velocity vector.
-    tof : float
-        Time of flight (s).
-
-    Note
-    ----
-    This method takes initial :math:`\vec{r}, \vec{v}`, calculates classical orbit parameters,
-    increases mean anomaly and performs inverse transformation to get final :math:`\vec{r}, \vec{v}`
-    The logic is based on formulae (4), (6) and (7) from http://dx.doi.org/10.1007/s10569-013-9476-9
-
-    """
-
-    # get the initial true anomaly and orbit parameters that are constant over time
-    p, ecc, inc, raan, argp, nu0 = rv2coe(k, r0, v0)
-
-    # get the initial mean anomaly
-    M0 = nu_to_M(nu0, ecc)
-    if np.abs(ecc - 1.0) > 1e-2:
-        # strong elliptic or strong hyperbolic orbits
-        a = p / (1.0 - ecc ** 2)
-        n = np.sqrt(k / np.abs(a ** 3))
-    else:
-        # near-parabolic orbit
-        q = p / np.abs(1.0 + ecc)
-        n = np.sqrt(k / 2.0 / (q ** 3))
-
-    M = M0 + tof * n
-    nu = M_to_nu(M, ecc)
-
-    return coe2rv(k, p, ecc, inc, raan, argp, nu)
 
 
 @jit
